LibJS: Change PropertyName(i32) => template<Integral T> PropertyName(T)
Negative numeric properties are not a thing (and we even VERIFY()'d this
in the constructor). It still allows using types with a negative range
for now as we have various places using int for example (without
actually needing the negative range, but that's a different story).

u32 is the internal type of `m_number` already, so this now allows us to
leverage the full u32 range for numeric properties.

+++ Userland/Libraries/LibJS/Runtime/ArrayPrototype.cpp
@@ -646,12 +646,12 @@ JS_DEFINE_NATIVE_FUNCTION(ArrayPrototype::slice)
     size_t index = 0;
 
     while (actual_start < final) {
-        bool present = this_object->has_property(actual_start);
+        bool present = this_object->has_property((u32)actual_start);
         if (vm.exception())
             return {};
 
         if (present) {
-            auto value = this_object->get(actual_start).value_or(js_undefined());
+            auto value = this_object->get((u32)actual_start).value_or(js_undefined());
             if (vm.exception())
                 return {};
